added Recurrences enum, added currentTripOffers

+++ engine/src/data/transpool/user/Feedbackable.java
@@ -1,5 +1,6 @@
 package data.transpool.user;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.collections.ObservableList;
 
@@ -7,6 +8,6 @@ public interface Feedbackable {
     void addFeedback(Feedback feedback);
     Feedback getFeedback(int riderID);
     ObservableList<Feedback> getAllFeedbacks();
-    int getAverageRating();
-    IntegerProperty averageRatingProperty();
+    double getAverageRating();
+    DoubleProperty averageRatingProperty();
 }
